@@ -6,12 +6,13 @@ use App\Models\ProductBrand;
 use Database\Seeders\ProductBrandSeeder;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
+use PHPUnit\Framework\Attributes\Test;
 
 class BrandGetTest extends TestCase
 {
   use RefreshDatabase;
 
-  /** @test */
+  #[Test]
   public function can_get_all_brands()
   {
     $this->seed(ProductBrandSeeder::class);
