@@ -6,22 +6,23 @@ use App\Models\ProductCategory;
 use Database\Seeders\ProductCategorySeeder;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
+use PHPUnit\Framework\Attributes\Test;
 
 class CategoryGetTest extends TestCase
 {
-  use RefreshDatabase;
+    use RefreshDatabase;
 
-  /** @test */
-  public function can_get_all_categories()
-  {
-    $this->seed(ProductCategorySeeder::class);
+    // #[Test]
+    public function test_can_get_all_categories(): void
+    {
+        $this->seed(ProductCategorySeeder::class);
 
-    $categories = ProductCategory::all();
+        $categories = ProductCategory::all();
 
-    $response = $this->getJson('/api/categories');
+        $response = $this->getJson('/api/categories');
 
-    $response->assertOk()
-      ->assertExactJson(['data' => $this->formatCategoryData($categories)])
-      ->assertJsonCount(3, 'data');
-  }
+        $response->assertOk()
+            ->assertExactJson(['data' => $this->formatCategoryData($categories)])
+            ->assertJsonCount(3, 'data');
+    }
 }
