@@ -4,299 +4,304 @@ namespace Tests\Feature\Api;
 
 use App\Models\Order;
 use App\Models\Product;
+use App\Models\ProductCategory;
 use Database\Seeders\ProductBrandSeeder;
 use Database\Seeders\ProductCategorySeeder;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Support\Arr;
 use Tests\TestCase;
+use PHPUnit\Framework\Attributes\Test;
 
 class ProductGetTest extends TestCase
 {
-  use RefreshDatabase;
+    use RefreshDatabase;
 
-  protected function setUp(): void
-  {
-    parent::setUp();
-    $this->seed([ProductCategorySeeder::class, ProductBrandSeeder::class]);
-  }
+    protected function setUp(): void
+    {
+        parent::setUp();
+        $this->seed([ProductCategorySeeder::class, ProductBrandSeeder::class]);
+    }
 
-  private function attemptToGetProductAndExpectOk(array $params = [])
-  {
-    $response = $this->getJson('/api/products?' . http_build_query($params));
+    private function attemptToGetProductAndExpectOk(array $params = [])
+    {
+        $response = $this->getJson('/api/products?' . http_build_query($params));
 
-    $response->assertOk()
-      ->assertJsonStructure([
-        'data' => [
-          '*' => [
-            'id',
-            'name',
-            'slug',
-            'image',
-            'category',
-            'brand',
-            'sex',
-            'price',
-            'stock',
-            'weight',
-            'sold_count',
-            'is_wishlist',
-          ],
-        ],
-        'page' => [
-          'total',
-          'per_page',
-          'current_page',
-          'last_page',
-          'from',
-          'to',
-        ],
-      ]);
+        $response->assertOk()
+            ->assertJsonStructure([
+                'data' => [
+                    '*' => [
+                        'id',
+                        'name',
+                        'slug',
+                        'image',
+                        'category',
+                        'brand',
+                        'sex',
+                        'price',
+                        'stock',
+                        'weight',
+                        'sold_count',
+                        'is_wishlist',
+                    ],
+                ],
+                'page' => [
+                    'total',
+                    'per_page',
+                    'current_page',
+                    'last_page',
+                    'from',
+                    'to',
+                ],
+            ]);
 
-    return $response;
-  }
+        return $response;
+    }
 
-  /** @test */
-  public function can_get_all_products()
-  {
-    $products = Product::factory(3)
-      ->hasImages()
-      ->create();
-    $this->createProduct(['is_publish' => false]);
+    // #[Test]
+    public function test_can_get_all_products()
+    {
 
-    $expectedProducts = $products->sortByDesc('id');
+        $products = Product::factory(3)
+            ->hasImages()
+            ->create();
+        $this->createProduct(['is_publish' => false]);
 
-    $response = $this->attemptToGetProductAndExpectOk();
+        $expectedProducts = $products->sortByDesc('id');
 
-    $response->assertJsonPath('data', $this->formatProductData($expectedProducts))
-      ->assertJsonCount(3, 'data');
+        $response = $this->attemptToGetProductAndExpectOk();
 
-    $this->assertStringStartsWith('http', $response['data'][0]['image']);
-  }
+        $response->assertJsonPath('data', $this->formatProductData($expectedProducts))
+            ->assertJsonCount(3, 'data');
 
-  /** @test */
-  public function can_get_products_by_page()
-  {
-    $this->createProduct(count: 23);
+        $this->assertStringStartsWith('http', $response['data'][0]['image']);
+    }
 
-    $response = $this->attemptToGetProductAndExpectOk(['page' => 2]);
+    // #[Test]
+    public function test_can_get_products_by_page()
+    {
+        $this->createProduct(count: 23);
 
-    $response->assertJsonPath('page', [
-      'total' => 23,
-      'per_page' => 20,
-      'current_page' => 2,
-      'last_page' => 2,
-      'from' => 21,
-      'to' => 23,
-    ])->assertJsonCount(3, 'data');
-  }
+        $response = $this->attemptToGetProductAndExpectOk(['page' => 2]);
 
-  /** @test */
-  public function can_get_products_by_search()
-  {
-    $products = Product::factory(3)
-      ->sequence(
-        [
-          'name' => 'Parfum Aroma Bunga Mawar',
-          'product_category_id' => 1,
-          'product_brand_id' => 1,
-        ],
-        [
-          'name' => 'Parfum Aroma Jeruk',
-          'product_category_id' => $this->createCategory(['name' => '100ml'])->id,
-          'product_brand_id' => 2,
-        ],
-        [
-          'name' => 'Parfum Wangi Bunga Jasmine',
-          'product_category_id' => 2,
-          'product_brand_id' => $this->createBrand(['name' => 'Roses Musk'])->id,
-        ],
-      )->create();
+        $response->assertJsonPath('page', [
+            'total' => 23,
+            'per_page' => 20,
+            'current_page' => 2,
+            'last_page' => 2,
+            'from' => 21,
+            'to' => 23,
+        ])->assertJsonCount(3, 'data');
+    }
 
-    // Search by name name
-    $response = $this->attemptToGetProductAndExpectOk(['search' => 'Bunga Mawar']);
+    #[Test]
+    public function can_get_products_by_search()
+    {
 
-    $response->assertJsonPath('data.0', $this->formatProductData($products[0]))
-      ->assertJsonCount(1, 'data');
 
-    // Search by category name
-    $response = $this->attemptToGetProductAndExpectOk(['search' => '100ml']);
+        $products = Product::factory(3)
+            ->sequence(
+                [
+                    'name' => 'Parfum Aroma Bunga Mawar',
+                    'product_category_id' => 1,
+                    'product_brand_id' => 1,
+                ],
+                [
+                    'name' => 'Parfum Aroma Jeruk',
+                    'product_category_id' => $this->createCategory(['name' => '100ml'])->id,
+                    'product_brand_id' => 2,
+                ],
+                [
+                    'name' => 'Parfum Wangi Bunga Jasmine',
+                    'product_category_id' => 2,
+                    'product_brand_id' => $this->createBrand(['name' => 'Roses Musk'])->id,
+                ],
+            )->create();
 
-    $response->assertJsonPath('data.0', $this->formatProductData($products[1]))
-      ->assertJsonCount(1, 'data');
+        // Search by name name
+        $response = $this->attemptToGetProductAndExpectOk(['search' => 'Bunga Mawar']);
 
-    // Search by brand name
-    $response = $this->attemptToGetProductAndExpectOk(['search' => 'Musk']);
+        $response->assertJsonPath('data.0', $this->formatProductData($products[0]))
+            ->assertJsonCount(1, 'data');
 
-    $response->assertJsonPath('data.0', $this->formatProductData($products[2]))
-      ->assertJsonCount(1, 'data');
-  }
+        // Search by category name
+        $response = $this->attemptToGetProductAndExpectOk(['search' => '100ml']);
 
-  /** @test */
-  public function can_sort_products_by_newest()
-  {
-    $products = $this->createProduct(count: 3);
+        $response->assertJsonPath('data.0', $this->formatProductData($products[1]))
+            ->assertJsonCount(1, 'data');
 
-    $expectedProducts = $products->sortByDesc('id');
+        // Search by brand name
+        $response = $this->attemptToGetProductAndExpectOk(['search' => 'Musk']);
 
-    $response =  $this->attemptToGetProductAndExpectOk(['sort_by' => 'newest']);
+        $response->assertJsonPath('data.0', $this->formatProductData($products[2]))
+            ->assertJsonCount(1, 'data');
+    }
 
-    $response->assertJsonPath('data', $this->formatProductData($expectedProducts))
-      ->assertJsonCount(3, 'data');
-  }
+    #[Test]
+    public function can_sort_products_by_newest()
+    {
+        $products = $this->createProduct(count: 3);
 
-  /** @test */
-  public function can_sort_products_by_oldest()
-  {
-    $products = $this->createProduct(count: 3);
+        $expectedProducts = $products->sortByDesc('id');
 
-    $expectedProducts = $products->sortBy('id');
+        $response = $this->attemptToGetProductAndExpectOk(['sort_by' => 'newest']);
 
-    $response =  $this->attemptToGetProductAndExpectOk(['sort_by' => 'oldest']);
+        $response->assertJsonPath('data', $this->formatProductData($expectedProducts))
+            ->assertJsonCount(3, 'data');
+    }
 
-    $response->assertJsonPath('data', $this->formatProductData($expectedProducts))
-      ->assertJsonCount(3, 'data');
-  }
+    #[Test]
+    public function can_sort_products_by_oldest()
+    {
+        $products = $this->createProduct(count: 3);
 
-  /** @test */
-  public function can_sort_products_by_sales()
-  {
-    $product1 = $this->createProductWithSales(quantities: [3, 2]);
-    $product2 = $this->createProductWithSales(quantities: [2, 1]);
-    $this->createProductWithSales(status: Order::STATUS_PENDING);
-    $this->createProductWithSales(status: Order::STATUS_PROCESSING);
-    $this->createProductWithSales(status: Order::STATUS_ON_DELIVERY);
+        $expectedProducts = $products->sortBy('id');
 
-    $response = $this->attemptToGetProductAndExpectOk(['sort_by' => 'sales']);
+        $response = $this->attemptToGetProductAndExpectOk(['sort_by' => 'oldest']);
 
-    $response
-      ->assertJsonPath('data.0.id', $product1->id)
-      ->assertJsonPath('data.0.sold_count', 5)
-      ->assertJsonPath('data.1.id', $product2->id)
-      ->assertJsonPath('data.1.sold_count', 3)
-      ->assertJsonCount(5, 'data');
+        $response->assertJsonPath('data', $this->formatProductData($expectedProducts))
+            ->assertJsonCount(3, 'data');
+    }
 
-    $this->assertCount(
-      3,
-      Arr::where($response['data'], fn ($product) => $product['sold_count'] === 0)
-    );
-  }
+    #[Test]
+    public function can_sort_products_by_sales()
+    {
+        $product1 = $this->createProductWithSales(quantities: [3, 2]);
+        $product2 = $this->createProductWithSales(quantities: [2, 1]);
+        $this->createProductWithSales(status: Order::STATUS_PENDING);
+        $this->createProductWithSales(status: Order::STATUS_PROCESSING);
+        $this->createProductWithSales(status: Order::STATUS_ON_DELIVERY);
 
-  /** @test */
-  public function can_sort_products_by_expensive()
-  {
-    $products = Product::factory(3)
-      ->sequence(
-        ['price' => 1000],
-        ['price' => 3000],
-        ['price' => 2000],
-      )
-      ->create();
+        $response = $this->attemptToGetProductAndExpectOk(['sort_by' => 'sales']);
 
-    $expectedProducts = $products->sortByDesc('price');
+        $response
+            ->assertJsonPath('data.0.id', $product1->id)
+            ->assertJsonPath('data.0.sold_count', 5)
+            ->assertJsonPath('data.1.id', $product2->id)
+            ->assertJsonPath('data.1.sold_count', 3)
+            ->assertJsonCount(5, 'data');
 
-    $response = $this->attemptToGetProductAndExpectOk(['sort_by' => 'expensive']);
+        $this->assertCount(
+            3,
+            Arr::where($response['data'], fn($product) => $product['sold_count'] === 0)
+        );
+    }
 
-    $response->assertJsonPath('data', $this->formatProductData($expectedProducts))
-      ->assertJsonCount(3, 'data');
-  }
+    #[Test]
+    public function can_sort_products_by_expensive()
+    {
+        $products = Product::factory(3)
+            ->sequence(
+                ['price' => 1000],
+                ['price' => 3000],
+                ['price' => 2000],
+            )
+            ->create();
 
-  /** @test */
-  public function can_sort_products_by_cheapest()
-  {
-    $products = Product::factory(3)
-      ->sequence(
-        ['price' => 3000],
-        ['price' => 1000],
-        ['price' => 2000]
-      )
-      ->create();
+        $expectedProducts = $products->sortByDesc('price');
 
-    $expectedProducts = $products->sortBy('price');
+        $response = $this->attemptToGetProductAndExpectOk(['sort_by' => 'expensive']);
 
-    $response = $this->attemptToGetProductAndExpectOk(['sort_by' => 'cheapest']);
+        $response->assertJsonPath('data', $this->formatProductData($expectedProducts))
+            ->assertJsonCount(3, 'data');
+    }
 
-    $response->assertJsonPath('data', $this->formatProductData($expectedProducts))
-      ->assertJsonCount(3, 'data');
-  }
+    #[Test]
+    public function can_sort_products_by_cheapest()
+    {
+        $products = Product::factory(3)
+            ->sequence(
+                ['price' => 3000],
+                ['price' => 1000],
+                ['price' => 2000]
+            )
+            ->create();
 
-  /** @test */
-  public function can_filter_products_by_category_id()
-  {
-    $products = Product::factory(3)
-      ->sequence(
-        ['product_category_id' => 1],
-        ['product_category_id' => 1],
-        ['product_category_id' => 2],
-      )
-      ->create();
+        $expectedProducts = $products->sortBy('price');
 
-    $expectedProducts = $products->where('product_category_id', 1)->sortByDesc('id');
+        $response = $this->attemptToGetProductAndExpectOk(['sort_by' => 'cheapest']);
 
-    $response = $this->attemptToGetProductAndExpectOk(['category_id' => 1]);
+        $response->assertJsonPath('data', $this->formatProductData($expectedProducts))
+            ->assertJsonCount(3, 'data');
+    }
 
-    $response->assertJsonPath('data', $this->formatProductData($expectedProducts))
-      ->assertJsonCount(2, 'data');
-  }
+    #[Test]
+    public function can_filter_products_by_category_id()
+    {
+        $products = Product::factory(3)
+            ->sequence(
+                ['product_category_id' => 1],
+                ['product_category_id' => 1],
+                ['product_category_id' => 2],
+            )
+            ->create();
 
-  /** @test */
-  public function can_filter_products_by_brand_id()
-  {
-    $products = Product::factory(3)
-      ->sequence(
-        ['product_brand_id' => 1],
-        ['product_brand_id' => 1],
-        ['product_brand_id' => 2],
-      )
-      ->create();
+        $expectedProducts = $products->where('product_category_id', 1)->sortByDesc('id');
 
-    $expectedProducts = $products->where('product_brand_id', 1)->sortByDesc('id');
+        $response = $this->attemptToGetProductAndExpectOk(['category_id' => 1]);
 
-    $response = $this->attemptToGetProductAndExpectOk(['brand_id' => 1]);
+        $response->assertJsonPath('data', $this->formatProductData($expectedProducts))
+            ->assertJsonCount(2, 'data');
+    }
 
-    $response->assertJsonPath('data', $this->formatProductData($expectedProducts))
-      ->assertJsonCount(2, 'data');
-  }
+    #[Test]
+    public function can_filter_products_by_brand_id()
+    {
+        $products = Product::factory(3)
+            ->sequence(
+                ['product_brand_id' => 1],
+                ['product_brand_id' => 1],
+                ['product_brand_id' => 2],
+            )
+            ->create();
 
-  /** @test */
-  public function can_filter_products_by_sex()
-  {
-    $products = Product::factory(3)
-      ->sequence(
-        ['sex' => 1],
-        ['sex' => 1],
-        ['sex' => 2],
-      )
-      ->create();
+        $expectedProducts = $products->where('product_brand_id', 1)->sortByDesc('id');
 
-    $expectedProducts = $products->where('sex', 1)->sortByDesc('id');
+        $response = $this->attemptToGetProductAndExpectOk(['brand_id' => 1]);
 
-    $response = $this->attemptToGetProductAndExpectOk(['sex' => 1]);
+        $response->assertJsonPath('data', $this->formatProductData($expectedProducts))
+            ->assertJsonCount(2, 'data');
+    }
 
-    $response->assertJsonPath('data', $this->formatProductData($expectedProducts))
-      ->assertJsonCount(2, 'data');
-  }
+    #[Test]
+    public function can_filter_products_by_sex()
+    {
+        $products = Product::factory(3)
+            ->sequence(
+                ['sex' => 1],
+                ['sex' => 1],
+                ['sex' => 2],
+            )
+            ->create();
 
-  /** @test */
-  public function can_filter_products_by_price_min_and_price_max()
-  {
-    $products = Product::factory(5)
-      ->sequence(
-        ['price' => 500],
-        ['price' => 3000],
-        ['price' => 1000],
-        ['price' => 5000],
-        ['price' => 7000],
-      )
-      ->create();
+        $expectedProducts = $products->where('sex', 1)->sortByDesc('id');
 
-    $expectedProducts = $products->whereBetween('price', [$min = 1000, $max = 5000])->sortByDesc('id');
+        $response = $this->attemptToGetProductAndExpectOk(['sex' => 1]);
 
-    $response = $this->attemptToGetProductAndExpectOk([
-      'price_min' => $min,
-      'price_max' => $max
-    ]);
+        $response->assertJsonPath('data', $this->formatProductData($expectedProducts))
+            ->assertJsonCount(2, 'data');
+    }
 
-    $response->assertJsonPath('data', $this->formatProductData($expectedProducts))
-      ->assertJsonCount(3, 'data');
-  }
+    #[Test]
+    public function can_filter_products_by_price_min_and_price_max()
+    {
+        $products = Product::factory(5)
+            ->sequence(
+                ['price' => 500],
+                ['price' => 3000],
+                ['price' => 1000],
+                ['price' => 5000],
+                ['price' => 7000],
+            )
+            ->create();
+
+        $expectedProducts = $products->whereBetween('price', [$min = 1000, $max = 5000])->sortByDesc('id');
+
+        $response = $this->attemptToGetProductAndExpectOk([
+            'price_min' => $min,
+            'price_max' => $max
+        ]);
+
+        $response->assertJsonPath('data', $this->formatProductData($expectedProducts))
+            ->assertJsonCount(3, 'data');
+    }
 }
