@@ -7,36 +7,36 @@ use Illuminate\Console\Command;
 
 class CancelExpiredOrder extends Command
 {
-  /**
-   * The name and signature of the console command.
-   *
-   * @var string
-   */
-  protected $signature = 'app:cancel-expired-order';
+    /**
+     * The name and signature of the console command.
+     *
+     * @var string
+     */
+    protected $signature = 'app:cancel-expired-order';
 
-  /**
-   * The console command description.
-   *
-   * @var string
-   */
-  protected $description = 'Command description';
+    /**
+     * The console command description.
+     *
+     * @var string
+     */
+    protected $description = 'Command description';
 
-  /**
-   * Execute the console command.
-   */
-  public function handle()
-  {
-    $orders = Order::where('status', Order::STATUS_PENDING_PAYMENT)
-      ->whereHas('invoice', fn ($q) => $q->where('due_date', '<', now()))
-      ->get();
+    /**
+     * Execute the console command.
+     */
+    public function handle()
+    {
+        $orders = Order::where('status', Order::STATUS_PENDING_PAYMENT)
+            ->whereHas('invoice', fn ($q) => $q->where('due_date', '<', now()))
+            ->get();
 
-    if ($orders->isNotEmpty()) {
-      foreach ($orders as $order) {
-        $order->update([
-          'status' => Order::STATUS_CANCELLED,
-          'cancel_reason' => 'Pesanan dibatalkan otomatis oleh sistem!'
-        ]);
-      }
+        if ($orders->isNotEmpty()) {
+            foreach ($orders as $order) {
+                $order->update([
+                    'status' => Order::STATUS_CANCELLED,
+                    'cancel_reason' => 'Pesanan dibatalkan otomatis oleh sistem!'
+                ]);
+            }
+        }
     }
-  }
 }
