@@ -4,29 +4,28 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
-  /**
-   * Run the migrations.
-   */
-  public function up(): void
-  {
-    Schema::create('payments', function (Blueprint $table) {
-      $table->id();
-      $table->foreignId('invoice_id')->cascadeOnDelete();
-      $table->enum('method', ['bank']);
-      $table->json('info');
-      $table->bigInteger('amount');
-      $table->enum('status', ['pending', 'cancelled', 'realeased']);
-      $table->timestamps();
-    });
-  }
+return new class extends Migration {
+    /**
+     * Run the migrations.
+     */
+    public function up(): void
+    {
+        Schema::create('payments', function (Blueprint $table) {
+            $table->id();
+            $table->foreignId('invoice_id')->cascadeOnDelete();
+            $table->enum('method', ['bank']);
+            $table->text('info');
+            $table->bigInteger('amount');
+            $table->enum('status', ['pending', 'cancelled', 'realeased']);
+            $table->timestamps();
+        });
+    }
 
-  /**
-   * Reverse the migrations.
-   */
-  public function down(): void
-  {
-    Schema::dropIfExists('payments');
-  }
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
+    {
+        Schema::dropIfExists('payments');
+    }
 };
