@@ -12,7 +12,7 @@ return new class extends Migration
             $table->id();
 
             $table->morphs('model');
-            $table->uuid('uuid')->nullable()->unique();
+            $table->uuid()->nullable()->unique();
             $table->string('collection_name');
             $table->string('name');
             $table->string('file_name');
