@@ -9,44 +9,38 @@ use App\Http\Controllers\Api\ProductController;
 use App\Http\Controllers\Api\RegionController;
 use App\Http\Controllers\Api\UserController;
 use App\Http\Controllers\Api\WishlistController;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
-/*
-|--------------------------------------------------------------------------
-| API Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register API routes for your application. These
-| routes are loaded by the RouteServiceProvider and all of them will
-| be assigned to the "api" middleware group. Make something great!
-|
-*/
+// Route::get('/user', function (Request $request) {
+//     return $request->user();
+// })->middleware('auth:sanctum');
 
 // Home Page
 Route::get('/home_page', HomePageController::class);
 
 // Product
 Route::controller(ProductController::class)->group(function () {
-  Route::get('/categories', 'categories');
-  Route::get('/brands', 'brands');
-  Route::get('/products', 'list');
-  Route::get('/products/{product}', 'get');
-  Route::get('/products/{product}/similars', 'similars');
+    Route::get('/categories', 'categories');
+    Route::get('/brands', 'brands');
+    Route::get('/products', 'list');
+    Route::get('/products/{product}', 'get');
+    Route::get('/products/{product}/similars', 'similars');
 });
 
 // Region
 Route::controller(RegionController::class)->group(function () {
-  Route::get('/region/provinces', 'provinces');
-  Route::get('/region/cities/{province}', 'cities');
+    Route::get('/region/provinces', 'provinces');
+    Route::get('/region/cities/{province}', 'cities');
 });
 
 // Auth
 Route::controller(AuthController::class)->group(function () {
-  Route::post('/auth/register', 'register');
-  Route::post('/auth/login', 'login');
-  Route::delete('/auth/logout', 'logout')->middleware('auth:sanctum');
-  Route::post('/auth/forgot_password', 'sendPasswordResetLink');
-  Route::put('/auth/reset_password', 'resetPassword');
+    Route::post('/auth/register', 'register');
+    Route::post('/auth/login', 'login');
+    Route::delete('/auth/logout', 'logout')->middleware('auth:sanctum');
+    Route::post('/auth/forgot_password', 'sendPasswordResetLink');
+    Route::put('/auth/reset_password', 'resetPassword');
 });
 
 // Shipping Cost
@@ -54,39 +48,39 @@ Route::post('/shipping_costs', [CheckoutController::class, 'shippingCosts']);
 
 Route::middleware(['auth:sanctum'])->group(function () {
 
-  // User Account
-  Route::controller(UserController::class)->group(function () {
-    Route::get('/user', 'get');
-    Route::put('/user', 'update');
-    Route::put('/user/change_password', 'updatePassword');
-  });
+    // User Account
+    Route::controller(UserController::class)->group(function () {
+        Route::get('/user', 'get');
+        Route::put('/user', 'update');
+        Route::put('/user/change_password', 'updatePassword');
+    });
 
-  // Checkout
-  Route::post('/checkout', [CheckoutController::class, 'checkout']);
+    // Checkout
+    Route::post('/checkout', [CheckoutController::class, 'checkout']);
 
-  // Order
-  Route::controller(OrderController::class)->group(function () {
-    Route::get('/orders', 'list');
-    Route::post('/orders', 'create');
-    Route::get('/orders/{order}', 'get');
-    Route::post('/orders/{order}/confirm_payment', 'confirmPayment');
-    Route::put('/orders/{order}/confirm_order_delivered', 'confirmDelivered');
-  });
+    // Order
+    Route::controller(OrderController::class)->group(function () {
+        Route::get('/orders', 'list');
+        Route::post('/orders', 'create');
+        Route::get('/orders/{order}', 'get');
+        Route::post('/orders/{order}/confirm_payment', 'confirmPayment');
+        Route::put('/orders/{order}/confirm_order_delivered', 'confirmDelivered');
+    });
 
-  // Wishlist
-  Route::controller(WishlistController::class)->group(function () {
-    Route::get('/wishlist', 'list');
-    Route::post('/wishlist', 'create');
-    Route::delete('/wishlist', 'delete');
-  });
+    // Wishlist
+    Route::controller(WishlistController::class)->group(function () {
+        Route::get('/wishlist', 'list');
+        Route::post('/wishlist', 'create');
+        Route::delete('/wishlist', 'delete');
+    });
 
-  // Cart
-  Route::controller(CartController::class)->group(function () {
-    Route::get('/carts', 'list');
-    Route::post('/carts', 'create');
-    Route::put('/carts/{cart}', 'update');
-    Route::delete('/carts', 'delete');
-  });
+    // Cart
+    Route::controller(CartController::class)->group(function () {
+        Route::get('/carts', 'list');
+        Route::post('/carts', 'create');
+        Route::put('/carts/{cart}', 'update');
+        Route::delete('/carts', 'delete');
+    });
 });
 
 // Route::fallback(fn () => abort(404));
