@@ -5,7 +5,7 @@ export default defineConfig({
     plugins: [
         laravel({
             input: [
-                "resources/sass/app.scss",
+                "resources/css/app.css",
                 "resources/js/app.js",
                 "resources/scss/adminlte.scss",
             ],
